+++ src/main/java/com/rflpazini/rogue/app/dataprovider/CrudPattern.java
@@ -1,4 +1,4 @@
-package com.rflpazini.rogue.dataprovider;
+package com.rflpazini.rogue.app.dataprovider;
 
 import java.util.List;
