+++ src/main/java/com/rflpazini/rogue/app/dataprovider/model/Money.java
@@ -1,4 +1,4 @@
-package com.rflpazini.rogue.dataprovider.model;
+package com.rflpazini.rogue.app.dataprovider.model;
 
 import lombok.Data;
