+++ src/test/java/com/rflpazini/rogue/domain/UserServiceTest.java
@@ -1,0 +1,6 @@
+package com.rflpazini.rogue.domain;
+
+
+public class UserServiceTest {
+
+}
