--- src/test/java/com/rflpazini/rogue/usecase/UserServiceTest.java
@@ -1,8 +1,0 @@
-package com.rflpazini.rogue.usecase;
-
-import org.junit.Test;
-
-
-public class UserServiceTest {
-
-}
